ofxget: work around lack of verbose arg in main parser
The problem is that the verbose arg contained in the add_subparser()
method, which is called for each ofxget command (i.e. where you
actually want the verbosity).  If you call ofxget with no command,
the namespace doesn't contain the arg, and it crashes rather than
displaying the help message.

This fix just adds a default verbosity to get over that.  We could
move the verbosity arg to the main parser, but it would never get
used without a command, so it's not terribly important.

Fixes #81