Preserve timezone when converting Time and DateTime to OFX.
Some OFX servers are particular about the formatting of the timezone in
`<DTSTART>` elements, and they reject the `[0:GMT]` timezone produced by
ofxtools.

Preserve the timezone of the `datetime` objects used when generating the OFX
timestamps to avoid this problem. Always format the UTX offset with a leading
`+` or `-`, even though the OFX spec permits the `+` to be dropped.

Move the timestamp formatting into a `format_datetime()` which is shared by the
`DateTime` and `Time` classes. This also ensures that millisecond rounding is
handled correctly for both.